refactor hepauv_hardware so we have timer_hardware, hepa_hardware, and led_hardware
