+++ hepa-uv/firmware/led_control_task/led_control_hardware.cpp
@@ -1,12 +1,9 @@
 #include "hepa-uv/firmware/led_control_hardware.hpp"
 
 #include "hepa-uv/core/led_control_task.hpp"
-#include "hepa-uv/firmware/hepauv_hardware.h"
+#include "hepa-uv/firmware/led_hardware.h"
 
 using namespace led_control_hardware;
-
-// NOLINTNEXTLINE(readability-convert-member-functions-to-static)
-auto LEDControlHardware::initialize() -> void { initialize_hepauv_hardware(); }
 
 void LEDControlHardware::set_button_led_power(uint8_t button, uint32_t r,
                                               uint32_t g, uint32_t b,
