+++ hepa-uv/firmware/timer_hardware.h
@@ -4,6 +4,8 @@
 
 #include "hepa-uv/core/constants.h"
 
+#include "platform_specific_hal_conf.h"
+
 // The frequency for the LED's pwm (2kHz)
 #define LED_PWM_FREQ_HZ (2000UL)
 // the frequency at which the timer should count so that it
@@ -14,14 +16,23 @@
 #define HEPA_PWM_FREQ_HZ (25000UL)
 #define HEPA_TIMER_FREQ (HEPA_PWM_FREQ_HZ * PWM_WIDTH)
 
-
 #ifdef __cplusplus
 extern "C" {
 #endif  // __cplusplus
 
-void set_button_led_pwm(PUSH_BUTTON_TYPE button, uint32_t red, uint32_t green, uint32_t blue, uint32_t white);
-void set_hepa_fan_pwm(uint32_t duty_cycle);
-void initialize_hepauv_hardware();
+extern TIM_HandleTypeDef htim1;
+extern TIM_HandleTypeDef htim3;
+extern TIM_HandleTypeDef htim8;
+extern TIM_HandleTypeDef htim16;
+extern TIM_HandleTypeDef htim20;
+
+extern TIM_OC_InitTypeDef htim1_sConfigOC;
+extern TIM_OC_InitTypeDef htim3_sConfigOC;
+extern TIM_OC_InitTypeDef htim8_sConfigOC;
+extern TIM_OC_InitTypeDef htim16_sConfigOC;
+extern TIM_OC_InitTypeDef htim20_sConfigOC;
+
+void initialize_pwm_hardware();
 
 #ifdef __cplusplus
 }  // extern "C"
