+++ include/hepa-uv/firmware/hepa_control_hardware.hpp
@@ -15,7 +15,6 @@ class HepaControlHardware : public HepaControlInterface {
     auto operator=(const HepaControlHardware&) -> HepaControlHardware& = delete;
     ~HepaControlHardware() final = default;
 
-    auto initialize() -> void;
     void set_hepa_fan_speed(uint32_t duty_cycle);
 };
