+++ include/hepa-uv/firmware/led_control_hardware.hpp
@@ -13,7 +13,6 @@ class LEDControlHardware : public led_control::LEDControlInterface {
     auto operator=(const LEDControlHardware&) -> LEDControlHardware& = delete;
     ~LEDControlHardware() final = default;
 
-    auto initialize() -> void;
     void set_button_led_power(uint8_t button, uint32_t r, uint32_t g,
                               uint32_t b, uint32_t w) final;
 };
